[SQUASH] Address review comments.
