+++ sycl/test/esimd/on-device/README
@@ -1,0 +1,4 @@
+This directory contains ESIMD tests which are run on Intel GPU device only.
+Some of them can run on host device too, but in general it is not always
+possible as some of ESIMD APIs (e.g. memory access via accessors) is not
+implemented for the host device.
